requirements.txt
fastapi
uvicorn
requests 
pytest
pyqrcode
pillow
pypng